backend/requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-generativeai>=0.3.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
